backend/requirements.txt
flask
psycopg2-binary
reportlab
prometheus_client
